[wip] renamed database package
