+++ tg-stonks-bot/bot_handlers_db_updates.py
@@ -3,15 +3,15 @@ from typing import Any
 from loguru import logger
 from pyrogram import Client as PyrogramClient
 
-from db import helpers
-from db.idatabase import IDatabase
+from database import helpers
+from database.protocol import IDatabase
 
 
 def _get_or_raise(the_kwargs: dict, key: str) -> Any:
     found = the_kwargs.get(key)
     if found is None:
         raise RuntimeError(
-            "Could not handle db updates:"
+            "Could not handle database updates:"
             f" argument '{key}' not specified"
             " to callback function"
         )
@@ -25,7 +25,7 @@ async def on_instrument_update(payload: dict, **kwargs):
     matched_tracking: list[dict] = idb.trackings_with({"instrument": instr_obj["id"]})
     for trk in matched_tracking:
         db_user_id = trk["tracked_by"]
-        logger.info(f"Inner db user id: {db_user_id}\n\t=> tracking: {trk}")
+        logger.info(f"Inner database user id: {db_user_id}\n\t=> tracking: {trk}")
 
         user_obj = idb.user_with({"id": db_user_id})
         user = db_helpers.to_user(user_obj)
