+++ tg-stonks-bot/database/errors.py
