+++ tg-stonks-bot/database/protocol.py
